+++ packages/chakra-ui-core/src/Accordion/tests/accordion.test.js
@@ -1,4 +1,4 @@
-import { Accordion, AccordionItem, AccordionHeader, AccordionPanel, AccordionIcon } from '@/packages/kiwi-core/src/Accordion'
+import { Accordion, AccordionItem, AccordionHeader, AccordionPanel, AccordionIcon } from '@/packages/chakra-ui-core/src/Accordion'
 import { render, userEvent, fireEvent } from '@/tests/test-utils'
 
 const renderComponent = (props) => {
